# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  apimodels.Event:
    properties:
      author:
        type: integer
      end:
        type: string
      id:
        type: integer
      roomId:
        type: integer
      start:
        type: string
      title:
        type: string
    type: object
...
  apimodels.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  apimodels.PatchEvent:
    properties:
      end:
        type: string
      roomId:
        type: integer
      start:
        type: string
    type: object
...
  apimodels.PostEvent:
    properties:
      end:
        type: string
      roomId:
        type: integer
      start:
        type: string
      title:
        type: string
    type: object
...
  apimodels.SuccessPostEvent:
    properties:
      id:
        type: integer
    type: object
...
info:
  contact: {}
  description: Swagger API for Dreamit-based web labs
  termsOfService: http://swagger.io/terms/
  title: Dreamit Swagger API
  version: "2.0"
paths:
...
  /events:
    get:
      description: Get events collection filtered by query parameters
      parameters:
      - description: Author ID
        in: query
        name: authorId
        type: integer
      - description: Room ID
        in: query
        name: roomId
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apimodels.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get events collection
      tags:
      - event
    post:
      consumes:
      - application/json
      description: Create event and get id
      parameters:
      - description: New event to add to the system
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/apimodels.PostEvent'
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apimodels.SuccessPostEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Post event
      tags:
      - event
  /events/{id}:
    delete:
      description: Delete event by id
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete event
      tags:
      - event
    get:
      description: Get event by id
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimodels.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get event
      tags:
      - event
    patch:
      consumes:
      - application/json
      description: Patch event roomId, start and end datetime by event id
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Patch editions to apply
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apimodels.PatchEvent'
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Reschedule event
      tags:
      - event
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"